requirements_no_audio.txt
pip
git+git://github.com/Rapptz/discord.py.git
youtube_dl
imgurpython
cleverbot
emoji
mcstatus
tweepy
pyfiglet
pillow
python-dateutil
tabulate
pytz
psutil
numpy
scipy
datadog
mutagen
bs4
feedparser